\name{plot.cortad}
\alias{plot.cortad}
\title{
Visualisation of CoRTAD time series product
}
\description{
Plots CoRTAD time series product
}
\usage{
plot.cortad(df, start = NULL, end = NULL, destfile = NULL, width = 8, height = 4)
}
\arguments{
  \item{df}{
a data frame created by get.cortad function
}
 \item{start}{
a character vector of the start date, e.g. "2013-01-01", default is NULL
}
 \item{end}{
a character vector of the end date, e.g. "2018-12-31",  default is NULL
}
  \item{destfile}{
a destination file to save the file, default is NULL
}
 \item{width}{
width of the plot to be saved; default is 8
}
 \item{height}{
height of the plot to be saved; default is 4
}
}
