\name{plot.crw}
\alias{plot.crw}
\title{
Visualistation of CRW time series product
}
\description{
Plots the CRW time series
}
\usage{
plot.crw(url, start = NULL, end = NULL, destfile = NULL, width = 8, height = 4)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{url}{
URL of the file to download from CRW database
}
 \item{start}{
a character vector of the start date, e.g. "2013-01-01", default is NULL
}
 \item{end}{
a character vector of the end date, e.g. "2018-12-31",  default is NULL
}
  \item{destfile}{
a destination file to save the file, default is NULL
}
 \item{width}{
width of the plot to be saved; default is 8
}
 \item{height}{
height of the plot to be saved; default is 4
}

}
