\name{plot.raster.cortad}
\alias{plot.raster.cortad}
\title{
Visualisation of extracted CoRTAD's raster
}
\description{
This function plots a raster object.
}
\usage{
plot.raster.cortad(r, ext)
}
\arguments{
  \item{r}{
a raster object
}
  \item{ext}{
a vector of four numbers representing the limits of x and y axes (i.e. xmin, xmax, ymin, ymax)
}
}
