\name{plot.the}
\alias{plot.the}
\title{
Visualisation of thermal stress metrics
}
\description{
Plots the thermal stress metrics
}
\usage{
plot.the(df, start = NULL, end = NULL, destfile = NULL, width = 8, height = 4)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{df}{
a data frame created by compute.the function
}
 \item{start}{
a character vector of the start date, e.g. "2013-01-01", default is NULL
}
 \item{end}{
a character vector of the end date, e.g. "2018-12-31",  default is NULL
}
  \item{destfile}{
a destination file to save the file, default is NULL
}
 \item{width}{
width of the plot to be saved; default is 8
}
 \item{height}{
height of the plot to be saved; default is 4
}
}
