\name{plot.tsm}
\alias{plot.tsm}
\title{
Visualisation of Thermal Stress Metrics
}
\description{
This function plots the thermal stress metrics.
}
\usage{
plot.tsm(x,destfile,width = 12, height = 8)
}
\arguments{
  \item{x}{
a data frame created by the function compute.tsm
}
  \item{destfile}{
a destination file to save the file, default (NULL)
}
 \item{width}{
width of the plot to be saved; default is 12
}
 \item{height}{
height of the plot to be saved; default is 8
}
}
